Revert "Bch replay protection"
