+++ src/common/plugin/types.ts
@@ -7,7 +7,6 @@ import {
 } from 'edge-core-js/types'
 
 import { IUTXO } from '../utxobased/db/types'
-import { ScriptTemplates } from '../utxobased/info/scriptTemplates/types'
 import { EngineEmitter } from './makeEngineEmitter'
 import { PluginState } from './pluginState'
 
@@ -51,7 +50,6 @@ export interface EngineCurrencyInfo extends EdgeCurrencyInfo {
   mempoolSpaceFeeInfoServer?: string
   customFeeSettings: CustomFeeSetting[]
   simpleFeeSettings: SimpleFeeSettings
-  scriptTemplates?: ScriptTemplates
 }
 
 export type CustomFeeSetting = 'satPerByte'
