+++ src/common/utxobased/info/bitcoincash.ts
@@ -1,6 +1,5 @@
 import { IMAGE_SERVER_URL } from '../../constants'
 import { EngineCurrencyInfo, EngineCurrencyType } from '../../plugin/types'
-import { BCHScriptTemplates } from './scriptTemplates/bitcoincashScriptTemplates'
 
 export const info: EngineCurrencyInfo = {
   currencyType: EngineCurrencyType.UTXO,
@@ -52,7 +51,5 @@ export const info: EngineCurrencyInfo = {
 
   // Images:
   symbolImage: `${IMAGE_SERVER_URL}/bitcoincash-logo-solo-64.png`,
-  symbolImageDarkMono: `${IMAGE_SERVER_URL}/bitcoincash-logo-solo-64.png`,
-
-  scriptTemplates: BCHScriptTemplates
+  symbolImageDarkMono: `${IMAGE_SERVER_URL}/bitcoincash-logo-solo-64.png`
 }
