--- src/common/utxobased/info/scriptTemplates/types.ts
@@ -1,14 +1,0 @@
-export interface ScriptTemplates {
-  [key: string]: ScriptTemplate
-}
-
-export type ScriptTemplate = (
-  pubkey: string,
-  optionalArgs?: OptionalScriptTemplateArgs
-) => string
-
-export interface OptionalScriptTemplateArgs {
-  cdsSig?: string
-  cdsMsg?: string
-  cdsPubKey?: string
-}
